int type for bool

+++ Modules/_functoolsmodule.c
@@ -49,7 +49,7 @@ typedef struct {
     PyObject *dict;        /* __dict__ */
     PyObject *weakreflist; /* List of weak references */
     vectorcallfunc vectorcall;
-    Py_ssize_t can_vcall;  /* Cache whether function allows vector call */
+    int can_vcall;         /* Cache whether function allows vector call */
 } partialobject;
 
 static void partial_setvectorcall(partialobject *pto);
